ebay_alerts_backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-enum-choices==2.1.3
django-rest-swagger==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.0
djangorestframework-simplejwt==5.0.0
ebaysdk==2.2.0
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
lxml==4.6.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
prompt-toolkit==3.0.22
psycopg2==2.9.2
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.1
requests==2.26.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0